refactor: split capital_choice port/road geometry into part file (#3290)
Phase 0 (C0) prerequisite slice for the colonizethis_logic package split:
decompose an over-500-line source file. Move the library-private port/road
placement geometry helpers (_topologyNodeById, _seaZonesAdjacentToProvince,
_isTileAdjacentToSea/SeaZone/OtherProvince, _nearestCoastalTileInProvince-
ForSeaZone, _shortestPathOnProvinceTiles) out of capital_choice.dart into a
new `capital_choice_port_road_geometry.dart` part file.

Pure structural extraction: no behavior change, public API unchanged (helpers
remain library-private via `part of`). capital_choice.dart drops 627 -> 463
physical lines (under the Phase 0 <=500-line target).

Refs #3290
